src/test/java/com/ljunggren/unitConversion/volume/unit/si/CubicCentimeterTest.java
package com.ljunggren.unitConversion.volume.unit.si;

import static org.junit.Assert.*;

import org.junit.Test;

import com.ljunggren.unitConversion.TestUnit;
import com.ljunggren.unitConversion.volume.Volume;
import com.ljunggren.unitConversion.volume.unit.si.CubicCentimeter;

public class CubicCentimeterTest implements TestUnit {
    
    private Volume volume = new CubicCentimeter();

    @Test
    public void toBaseUnitTest() {
        assertEquals(1, volume.toBaseUnit().apply(1000.0), 0);
        assertEquals(0, volume.toBaseUnit().apply(0.0), 0);
    }

    @Test
    public void fromBaseUnitTest() {
        assertEquals(1000, volume.fromBaseUnit().apply(1.0), 0);
        assertEquals(0, volume.fromBaseUnit().apply(0.0), 0);
    }

    @Test
    public void systemTest() {
        assertNotNull(volume.system());
    }

    @Test
    public void labelTest() {
        assertNotNull(volume.label());
    }

}
